requirements.txt
fastapi==0.111.0
pinecone-client~=2.2.2
sqlalchemy~=2.0
alembic-postgresql-enum==1.1.0
alembic~=1.13
loguru==0.7.2
python-dotenv
pydantic-settings
llama-index-embeddings-gemini
pyjwt
loguru
aioboto3
asyncpg
psycopg2
redis
aioredis==2.0.1
openai